pyproject.toml
[project]
name = "sram-forge"
version = "0.1.0"
description = "Generate SRAM-based chip designs for GF180MCU"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-PyYAML>=6.0",
]

[project.scripts]
sram-forge = "sram_forge.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sram_forge"]

[tool.pytest.ini_options]
testpaths = ["sram_forge/tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
exclude = ["project-template"]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.isort]
known-first-party = ["sram_forge"]

[tool.mypy]
python_version = "3.11"
warn_unused_configs = true
disallow_untyped_defs = false  # Start permissive, tighten over time
check_untyped_defs = true
# Permissive settings for initial setup - tighten over time
warn_return_any = false  # Jinja render() returns Any
strict = false

[[tool.mypy.overrides]]
module = "sram_forge.tests.*"
# Tests intentionally pass dicts to Pydantic models for coercion testing
disable_error_code = ["arg-type"]

[[tool.mypy.overrides]]
module = "sram_forge.models.chip"
# Literal[float] not supported by mypy, but works at runtime with Pydantic
disable_error_code = ["valid-type"]

[[tool.mypy.overrides]]
module = "sram_forge.cli.main"
# Import fallback pattern for __version__ causes redefinition and import-not-found warnings
disable_error_code = ["no-redef", "import-not-found"]
